PI-276 Initial build, test, push CI workflow
The workflow is triggered on a push to any branch. On a non-main branch, the workflow will stop after the Gradle build/test step by default, but this can be overridden in the UI (via workflow_dispatch)

In the end, I couldn't rely on the Gradle build cache for determining which images needed to be built and deployed, so had to add a custom "check-changes" action for parsing the git diff.