PI-406 Send failed Logstash events to the SQS DLQ
This commit updates the Logstash pipelines to handle two main types of failures:
1. failures in the `filter` plugins (e.g. JSON parsing)
2. failures in the `output` plugins (e.g. non-retryable error responses from Elasticsearch/OpenSearch)

Note that failures in the `input` plugins will result in the pipeline being terminated.

### `filter` failures
Typically, filters add "tags" to the event when an error occurs - rather than actually failing or dropping the event. This is problematic as we could accidentally index partial/invalid documents, if for example the JDBC lookup fails in the incremental pipeline.

This is now handled by checking for the presence of tags and sending the message to the SQS DLQ.

### `output` failures
The `elasticsearch`/`opensearch` output plugins will retry most errors indefinitely, however 400 and 404 errors will be sent to an internal Logstash DLQ. Two additional "dlq" pipelines have been added to capture these and push them to our SQS DLQ.

Note: 409 errors (i.e. version mismatches) will always be logged and dropped, but we shouldn't get any of these as we don't use document versioning.