PI-426 Use shared task definition values from templates/defaults.yml (#163)
Moves cross-project deployment values into a shared defaults.yml file. This allows us to specify a common CPU/Memory footprint, that can be overridden easily per-project or per-environment.

It also switches to using soft/hard limits for CPU and Memory, so that only a small amount (0.125CPU x 128MB) is reserved on the cluster by default, with the ability to burst as needed (to 1CPU x 1GB).