PI-463 Refactor to handle messages without eventType
This updates our `MessageListener`s to take the full Notification object, including the message attributes.  This means whenever the `eventType` is not present in the message itself, we can still use it from the attributes in the listener.  It also enables us to use any other attributes as needed in future.