Add multiplatform Eclipse Temurin image
After trying in https://github.com/ministryofjustice/community-api/pull/732
to switch to a multiplatform build, it became apparent that there are
no arm64 17 jre alpine Docker images, and also the underlying image in
this repo is not multiplatform. This adds a new multiplatform image
based on the version 19 Eclipse Temurin image, which is multiplatform.