ApprovedPremises-and-Delius: changes to event details (#680)
* Extend submittedBy in application.submitted event

Rather than bearing a representation of a `staffMember` it now carries
representations of various geographic entities as well as the submitting
staff member:

```
"submittedBy": {
  "staffMember": {
    "staffCode": "N54A999",
    "staffIdentifier": 1501234567,
    "forenames": "John",
    "surname": "Smith",
    "username": "JohnSmithNPS"
  },
  "probationArea": {
    "code": "N02",
    "name": "NPS North East"
  },
  "cru": {
    "code": "N02",
    "name": "NPS North East"
  },
  "team": {
    "code": "N54NGH",
    "name": "Gateshead 1"
  },
  "ldu": {
    "code": "N54PPU",
    "name": "Public Protection NE"
  },
  "region": {
    "code": "NE",
    "name": "North East"
  }
}
```

* Move probationArea into submittedBy in application.submitted

This is the probation area of the applicant and so to avoid ambiguity belongs
within the `SubmittedBy` entity.

* Extend assessedBy key in application.assessed

Rather than just containing a staff member, it now consists of 3 objects:

- cru
- probationArea
- staffMember

i.e. we want to make it clear that these are associated with the assessment
step rather than any other part of the AP journey.

* Extend bookedBy key in booking.made event

The key now contains two objects:

- staffMember
- CRU

which we want to explicitly associate with the booking step as distinct
from any other part of the AP journey.