PI-830 Added contact search indexing pipeline (#942)
## Changes
* Added contact-full-load and contact-incremental pipelines
* Added support for conditionally enabling pipelines using the `PIPELINES_ENABLED` environment variable
* Consolidated DLQ pipelines (used for pushing Logstash DLQ to SQS DLQ)
* Switched full-load pipelines to use a continuous schedule, instead of Logstash jdbc_paging
* Use "lastId", instead of "count", for indicating when a full re-index has completed
* Updated cron jobs to perform re-indexing and alias switching in the same container
* Added Sentry error reporting
* Removed single pod limit and reduced resources for Kubernetes deployment (it only needs to handle incremental updates now, the full load happens in the cron job)
* Removed "management" image
* Added support for local testing with docker-compose and LocalStack/SQS
* Expose Logstash API for monitoring pipeline stats - e.g. https://person-search-index-from-delius-preprod.hmpps.service.justice.gov.uk/_node/stats/pipelines
* Re-indexing can now be triggered manually - e.g. for contacts in preprod:
```shell
kubectl -n hmpps-probation-integration-services-preprod create job --from=cronjob/contact-reindex test-contact-reindex
kubectl -n hmpps-probation-integration-services-preprod logs -f job/test-contact-reindex
```

## Issues addressed
The following issues were identified that prevented the current pattern from working nicely with the contact table.  They've been addressed by the changes listed above.

### Row count metadata document (resolved)
Passing the **row count** as a document to let the cronjob know when to switch aliases won't work for contacts, even getting a count of the contact table takes too long (hours).

Resolved by switching to use the **max id**, and indexing in order of id, then waiting for a document with that ID to be indexed.

### Gaps in the contact table (resolved)
The Logstash [jdbc_paging](https://www.elastic.co/guide/en/logstash/8.6/plugins-inputs-jdbc.html#plugins-inputs-jdbc-jdbc_paging_enabled) mechanism doesn't handle gaps between rows very well, because it blindly adds `:size` to `:offset` with each query execution.  This means, if a gap exists that is significantly larger than the page size, it will have to run lots of queries before it gets to the next row.

For example: if there's a gap of 200 rows (say between id=1 and id=200) and we set the page size to 10, then Logstash will run a query for each of the 20 pages it takes to bridge the gap.
```
id:     --1-------------------200--
offset: --1--10--20--30--...--200--
```

Here are a few examples of the big gaps in the contact table
| contact_id min | contact_id max | gap width      |
|----------------|----------------|----------------|
| 5              | 10001898       | 10,001,893     |
| 14886644       | 40063018       | 25,176,374     |
| 42625562       | 70011485       | 27,385,923     |
| ...            | ...            | ...            |
| 1823322687     | 90000000004    | 88,176,677,317 |

This means with a page size of 100,000, Logstash would run 881,766 queries to bridge the final gap.  If each query took only 100ms, that would still add an extra 24 hours to the total execution time.  So, unfortunately, jdbc_paging is not feasible for this.

Resolved by switching to manual batching in the query, making use of the `:sql_last_run` parameter and a continuous schedule (`* * * * *`).  This also enables us to schedule the re-indexing using Kubernetes, rather than relying on Logstash scheduling.