PI-1093 Reject and do not reprocess AppointmentNotFound errors
If we cannot find the Delius contact handling feedback for an appointment, we no longer add the message to the DLQ. Instead, we log the reason to Telemetry and drop the message. The reason is usually that the NSI container has been deleted or terminated, resulting in the contacts also being deleted.