PI-1130 Increase batch frequency and reduce batch size
Previously the person-reindex job was trying to fetch too much from the database, and only ran the query once per-minute. This meant that, towards the end of the indexing process, by the time the SQL finished running more records had been created in the database - so it got stuck in a cat-and-mouse loop of trying to get to the end of the offender table.

The reason for fetching so much from the database was to ensure that there were enough records to keep Logstash busy for a full minute until the next query ran, to avoid quiet periods.  However, Logstash supports seconds in its cron syntax - so we can make this more efficient by setting a smaller batch size and running more often. This also means queries will complete in a few 100ms, rather than a few minutes, which drastically reduces the chance of records being created while the queries are still running.

I've tested this in preprod, and it works as expected - throughput is more consistent, and doesn't get stuck at the end.

This PR also sets the CPU/Memory requests to the observed values from the running cron jobs.