PI-1558 Update PSR extension to PDF when complete
Previously, an empty PDF would be uploaded to the Alfresco when the initial draft was started.  Following the recent changes, Delius now uploads a .doc file instead as the draft.  However, the filename remains the same after the report is completed.

So currently we end up with a PDF document saved as e.g.
```
Nat Short format pre-sentence report_13102023_112036_Larson_D_X729742.doc
```

This PR updates the extension to .pdf before saving