Seed probation case for Approved Premises E2E testing
The Approved Premises local E2E tests have need of a "probation case"
(person leaving custody) to use in E2E tests which rely on a manual or
"ad hoc" booking. These bookings can only be made for people who don't
have any existing applications with the service. This is most reliably
achieved by using a distinct probation case.

In this commit we seed a case with CRN X320811, which [ties in with
seeding at CommunityAPI][] -- which is still being used at present for fetching
a list of convictions:

```
http://community-api/secure/offenders/crn/X320811/convictions
```

[ties in with seeding at CommunityAPI]:
https://github.com/ministryofjustice/community-api/blob/736ac63ff7705a0c8d998ed2773d59563b0fdea4/src/main/resources/db/data/V1_9__offender_X320811_custody_status_data.sql#L1