PI-2676 Don't create a release if there is already an un-recalled release (#4482)
The integration service should never create a release where there is already an active release.  Delius expects releases and recalls to be in order and alternating, so you would never have two releases in a row, unless there was a recall between them.

Previously the `releaseDateValid` logic didn't account for an un-recalled release.  If `custody.mostRecentRelease()` is not null, but `custody.mostRecentRelease().recall` _is_ null, then it returned `true` but it should have returned `false`.