PI-2526 Switch to cosine similarity + increase chunk size (#4552)
The Data Science team have been using cosing similarity for their testing, whereas OpenSearch defaults to L2 (Euclidean distance).

Also increasing chunk size as a workaround for the batch size limit of 32 in the mxbai model, which limits the length of the notes we can process in one go to 1792 chars (now 3584).