PI-2930 Remove duplicate key dates (#4836)
* PI-2930 Remove duplicate key dates

Previously, when we encountered a case that had two key dates of the same type (with different dates), we would try to update the first one we found.  However, there's a chance we update the date to be the same as the date on the other key date - resulting in a key constraint error.  With this change, if we come across a case with multiple key dates of the same type - we delete all but one, then update the remaining one.

Resolves https://ministryofjustice.sentry.io/issues/6279152215

* Ensure deletes happen before updates