PI-3729 Log and ignore missing offence codes (#5903)
The main change here is to look up the Home Office code from the Manage
Offences API at the start, and to filter out any offences with a null or
ignored CJA or Home Office code, so that we don't have to deal with them
again later. Previously, we were using the Home Office code from two
places (Manage Offences API + Detailed Offence table) - I've made the
assumption that the API is the better source for this, but it would be
good if someone could confirm.

Also, we were previously only applying the offence code ignore rules to
the main offence - by applying these rules to all offences at the start,
we can ensure additional offences with invalid codes don't slip through.

Depends on #5902

Resolves
[COMMON-PLATFORM-AND-DELIUS-W](https://ministryofjustice.sentry.io/issues/6689394978)
and
[COMMON-PLATFORM-AND-DELIUS-1R](https://ministryofjustice.sentry.io/issues/6859282019)